+++ services/madoc-ts/src/frontend/shared/utility/scroll-top.tsx
@@ -5,7 +5,15 @@ function _ScrollTop({ history }: any) {
   useEffect(() => {
     const listener = history.listen(() => {
       if (typeof window !== 'undefined') {
-        window.scrollTo(0, 0);
+        const hash = window.location.hash.slice(1);
+        if (!hash) {
+          window.scrollTo(0, 0);
+        } else {
+          const $el = document.getElementById(hash);
+          if ($el) {
+            $el.scrollIntoView({ block: 'start' });
+          }
+        }
       }
     });
     return () => {